pyproject.toml
[project]
name = "youtube-daily-comments-intelligence"
version = "1.0.0"
description = "Daily YouTube comments intelligence for playlist-driven news channels."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlalchemy>=2.0.36",
  "psycopg[binary]>=3.2.3",
  "alembic>=1.14.0",
  "pydantic-settings>=2.6.1",
  "celery[redis]>=5.4.0",
  "redis>=5.2.1",
  "httpx>=0.28.1",
  "requests>=2.32.0",
  "python-multipart>=0.0.20",
  "tenacity>=9.0.0",
  "openai>=1.59.9",
  "numpy>=2.2.1",
  "scikit-learn>=1.6.0",
  "hdbscan>=0.8.40",
  "sentence-transformers>=3.3.1",
  "langdetect>=1.0.9",
  "isodate>=0.7.2",
  "markdown>=3.7",
  "jinja2>=3.1.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "ruff>=0.8.6",
  "black>=24.10.0",
  "mypy>=1.14.1",
  "pre-commit>=4.0.1",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["alembic/versions"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4"]
ignore = ["E501"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Body", "fastapi.Depends", "fastapi.Query"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = ["alembic"]
